refactor(frontend): decouple wallet logic into useWalletManager hook
